refactor: move util functions to `utils`
